refactor(game): move reducers into lib/state (T28)
Relocate the four state-layer modules (create-initial-state,
update-player, update-enemies, move-enemy-toward-player) and their
colocated tests into packages/game/lib/state/. Imports of core/
modules become ../core/...; the only outside importer was lib/loop.ts,
updated to ./state/...
