+++ packages/game/lib/state/update-player.test.ts
@@ -1,7 +1,7 @@
 import { brand, createPrng } from "@bruff/utils";
 import { describe, expect, it } from "vitest";
-import { PLAYER_SIZE, PLAYER_SPEED } from "./core/constants.js";
-import type { GameState } from "./core/types.ts";
+import { PLAYER_SIZE, PLAYER_SPEED } from "../core/constants.js";
+import type { GameState } from "../core/types.ts";
 import updatePlayer from "./update-player.js";
 
 const ZERO = 0;
