+++ packages/game/lib/state/update-player.ts
@@ -1,7 +1,7 @@
 /* eslint-disable unicorn/switch-case-braces */
-import { PLAYER_SIZE, PLAYER_SPEED, ZERO } from "./core/constants.js";
+import { PLAYER_SIZE, PLAYER_SPEED, ZERO } from "../core/constants.js";
 import { clamp } from "@bruff/utils";
-import type { GameState } from "./core/types.ts";
+import type { GameState } from "../core/types.ts";
 
 const getMovementDelta = (
   key: string | undefined,
